refactor: move isMobile useState handler into a separate global helper function.
