+++ src/utils/responsivenessHelpers.ts
@@ -1,0 +1,21 @@
+
+import { useEffect, useState } from "react";
+
+const isMobileWindowSize = () => {
+    const [isMobile, setIsMobile] = useState(false);
+
+    useEffect(() => {
+        const mediumWindowSize = 768;
+        const handleResize = () => {
+            setIsMobile(window.innerWidth < mediumWindowSize);
+        };
+
+        handleResize();
+        window.addEventListener("resize", handleResize);
+        return () => window.removeEventListener("resize", handleResize);
+    }, []);
+
+    return isMobile
+}
+
+export default isMobileWindowSize;
